requirements.txt
algoliasearch-django
django
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
pillow
autopep8
django-rest-auth
django-allauth
django-mptt